Change entry point. Tests still not working.
