+++ app/gui/index.html
@@ -50,10 +50,10 @@
   </head>
   <body>
     <div id="enso-spotlight" class="enso-spotlight"></div>
-    <div id="enso-dashboard" class="enso-dashboard"></div>
+    <div id="enso-app" class="enso-app"></div>
     <div id="enso-chat" class="enso-chat"></div>
     <div id="enso-portal-root" class="enso-portal-root"></div>
-    <script type="module" src="/src/entrypoint.ts"></script>
+    <script type="module" src="/src/index.ts"></script>
     <noscript> This page requires JavaScript to run. Please enable it in your browser. </noscript>
   </body>
 </html>
