+++ app/gui/src/dashboard/components/AriaComponents/Dialog/utilities.ts
@@ -18,7 +18,7 @@ const IGNORE_INTERACT_OUTSIDE_ELEMENTS = [
   // ReactQuery devtools
   '.tsqd-parent-container',
   // Our components that should ignore the interact outside event
-  ':is(.enso-dashboard, .enso-chat, .enso-portal-root) [data-ignore-click-outside]',
+  ':is(.enso-app, .enso-chat, .enso-portal-root) [data-ignore-click-outside]',
 ]
 
 const IGNORE_INTERACT_OUTSIDE_ELEMENTS_SELECTOR = `:is(${IGNORE_INTERACT_OUTSIDE_ELEMENTS.join(', ')})`
