+++ app/gui/src/dashboard/styles.css
@@ -20,6 +20,6 @@
   }
 }
 
-:where(.enso-dashboard) {
+:where(.enso-app, .App) {
   @apply absolute inset-0 isolate flex flex-col overflow-hidden;
 }
