--- app/gui/src/project-view/asyncApp.ts
@@ -1,7 +1,0 @@
-import '@/assets/base.css'
-
-/** Load App.vue asynchronously. */
-export async function AsyncApp() {
-  const app = await import('@/App.vue')
-  return app
-}
