+++ app/gui/src/project-view/providers/eventLogging.ts
@@ -1,12 +1,13 @@
-import { computed, type Ref } from 'vue'
+import { ToValue } from '@/util/reactivity'
+import { computed, Ref, toValue } from 'vue'
 import { createContextStore } from '.'
 
 export type EventLogger = ReturnType<typeof injectEventLogger>
 export const [provideEventLogger, injectEventLogger] = createContextStore(
   'event logger',
-  (logEvent: Ref<LogEvent>, projectId: Ref<string>) => {
+  (logEvent: Ref<LogEvent>, projectId: ToValue<string>) => {
     const logProjectId = computed(() => {
-      const id = projectId.value
+      const id = toValue(projectId)
       if (!id) return undefined
       const prefix = 'project-'
       const projectUuid = id.startsWith(prefix) ? id.substring(prefix.length) : id
