+++ package.json
@@ -40,8 +40,8 @@
     "ci:chromatic:react": "corepack pnpm run -r --filter enso-gui chromatic:react",
     "ci:chromatic:vue": "corepack pnpm run -r --filter enso-gui chromatic:vue"
   },
+  "//": "To completely ignore deep dependencies, see .pnpmfile.cjs",
   "pnpm": {
-    "//": "To completely ignore deep dependencies, see .pnpmfile.cjs",
     "overrides": {
       "tslib": "$tslib",
       "jsdom": "^24.1.0",
